+++ server/models/word.js
@@ -1,41 +1,41 @@
-'use strict';
-const {
-  Model
-} = require('sequelize');
-module.exports = (sequelize, DataTypes) => {
+import { Model } from "sequelize";
+export default (sequelize, DataTypes) => {
   class Word extends Model {
     static associate(models) {
       Word.hasMany(models.UserWord, {
-        foreignKey: 'idWords',
-        as: 'userWords'
+        foreignKey: "idWords",
+        as: "userWords",
       });
     }
   }
-  Word.init({
-    word: {
-      type: DataTypes.STRING,
-      allowNull: false,
-      unique: true,
-      validate: {
-        notEmpty: true,
-      }
-    },
-    hint: {
-      type: DataTypes.STRING,
-      allowNull: false,
-      validate: {
-        notEmpty: true,
-        is: /^[A-zÀ-ú]+$/
+  Word.init(
+    {
+      word: {
+        type: DataTypes.STRING,
+        allowNull: false,
+        unique: true,
+        validate: {
+          notEmpty: true,
+        },
+      },
+      hint: {
+        type: DataTypes.STRING,
+        allowNull: false,
+        validate: {
+          notEmpty: true,
+          is: /^[A-zÀ-ú]+$/,
+        },
+      },
+      description: {
+        type: DataTypes.STRING,
+        allowNull: true,
       },
     },
-    description: {
-      type: DataTypes.STRING,
-      allowNull: true,
+    {
+      sequelize,
+      modelName: "Word",
+      raw: true,
     }
-  }, {
-    sequelize,
-    modelName: 'Word',
-    raw: true,
-  });
+  );
   return Word;
 };
